moved multi-gpu helper functions to utils
